+++ packages/google_maps_flutter/google_maps_flutter_web/example/integration_test/google_maps_controller_test.dart
@@ -21,6 +21,9 @@ import 'google_maps_controller_test.mocks.dart';
   MockSpec<PolygonsController>(onMissingStub: OnMissingStub.returnDefault),
   MockSpec<PolylinesController>(onMissingStub: OnMissingStub.returnDefault),
   MockSpec<MarkersController>(onMissingStub: OnMissingStub.returnDefault),
+  MockSpec<html.Geolocation>(onMissingStub: OnMissingStub.returnDefault),
+  MockSpec<html.Geoposition>(onMissingStub: OnMissingStub.returnDefault),
+  MockSpec<html.Coordinates>(onMissingStub: OnMissingStub.returnDefault),
 ])
 
 /// Test Google Map Controller
@@ -481,6 +484,107 @@ void main() {
           controller.debugSetOverrides(createMap: (_, __) => map);
           controller.init();
           expect(controller.trafficLayer, isNotNull);
+        });
+      });
+
+      group('My Location', () {
+        testWidgets('by default is disabled', (WidgetTester tester) async {
+          controller = createController();
+          controller.init();
+
+          expect(controller.myLocationButton, isNull);
+        });
+
+        testWidgets('initializes with my location & display my location button',
+            (WidgetTester tester) async {
+          late final MockGeolocation mockGeolocation = MockGeolocation();
+          late final MockGeoposition mockGeoposition = MockGeoposition();
+          late final MockCoordinates mockCoordinates = MockCoordinates();
+          const LatLng currentLocation = LatLng(10.8231, 106.6297);
+
+          controller = createController(
+              mapConfiguration: const MapConfiguration(
+            myLocationEnabled: true,
+            myLocationButtonEnabled: true,
+          ));
+
+          controller.debugSetOverrides(
+            createMap: (_, __) => map,
+            markers: markers,
+            geolocation: mockGeolocation,
+          );
+
+          when(mockGeoposition.coords).thenReturn(mockCoordinates);
+
+          when(mockCoordinates.longitude).thenReturn(currentLocation.longitude);
+
+          when(mockCoordinates.latitude).thenReturn(currentLocation.latitude);
+
+          when(mockGeolocation.getCurrentPosition(
+            timeout: anyNamed('timeout'),
+          )).thenAnswer((_) async => mockGeoposition);
+
+          when(mockGeolocation.watchPosition()).thenAnswer((_) {
+            return Stream<MockGeoposition>.fromIterable(
+                <MockGeoposition>[mockGeoposition]);
+          });
+
+          controller.init();
+          await tester.pumpAndSettle();
+
+          final Set<Marker> capturedMarkers =
+              verify(markers.addMarkers(captureAny)).captured[1] as Set<Marker>;
+
+          expect(controller.myLocationButton, isNotNull);
+          expect(capturedMarkers.length, 1);
+          expect(capturedMarkers.first.position, currentLocation);
+          expect(capturedMarkers.first.zIndex, 0.5);
+        });
+
+        testWidgets('initializes with my location only',
+            (WidgetTester tester) async {
+          late final MockGeolocation mockGeolocation = MockGeolocation();
+          late final MockGeoposition mockGeoposition = MockGeoposition();
+          late final MockCoordinates mockCoordinates = MockCoordinates();
+          const LatLng currentLocation = LatLng(10.8231, 106.6297);
+
+          controller = createController(
+              mapConfiguration: const MapConfiguration(
+            myLocationEnabled: true,
+            myLocationButtonEnabled: false,
+          ));
+          controller.debugSetOverrides(
+            createMap: (_, __) => map,
+            markers: markers,
+            geolocation: mockGeolocation,
+          );
+
+          when(mockGeoposition.coords).thenReturn(mockCoordinates);
+
+          when(mockCoordinates.longitude).thenReturn(currentLocation.longitude);
+
+          when(mockCoordinates.latitude).thenReturn(currentLocation.latitude);
+
+          when(mockGeolocation.getCurrentPosition.call(
+            timeout: const Duration(seconds: 30),
+          )).thenAnswer((_) async => mockGeoposition);
+
+          when(mockGeolocation.watchPosition()).thenAnswer((_) {
+            return Stream<MockGeoposition>.fromIterable(
+                <MockGeoposition>[mockGeoposition]);
+          });
+
+          controller.init();
+
+          await tester.pumpAndSettle();
+
+          final Set<Marker> capturedMarkers =
+              verify(markers.addMarkers(captureAny)).captured[1] as Set<Marker>;
+
+          expect(controller.myLocationButton, isNull);
+          expect(capturedMarkers.length, 1);
+          expect(capturedMarkers.first.position, currentLocation);
+          expect(capturedMarkers.first.zIndex, 0.5);
         });
       });
     });